Clean up Gutenberg layout file
